Unfinished -- RGB based Combat System
Unfinished -- working on this in my spare time. This is the basic framework for an RGB based combat system. As players kill enemies of matching colors, the kill counter will go up. If the player kills and enemy of an off-color, they players color and damage type will change, and the counter will reset.