November Game Jam -- THEME: Parasite and Candle
Basic scripts for a game jam submission from November 2022. The theme was to build a game centered around a Parasite, and to use a Candle somehow in the game. My submission was a maze-runner, where the player had to carry a parasite through the maze. The parasite would slowly drain the players health, and when the player used the candle to release the parasite, the parasite would slowly lose health. Player could regain health by picking up food hidden in the mazes.

The balancing act of carrying the parasite and not leaving it for too long created a frantic and interesting pace for the player.